08._Server-Server/02._python/pyproject.toml
[tool.poetry]
name = "02-python"
version = "0.1.0"
description = ""
authors = ["doro-hd <doro-hd@protonmail.com>"]
readme = "README.md"
packages = [{include = "02"}]

[tool.poetry.dependencies]
python = "^3.13"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
